+++ src/versioning/Versioning.ts
@@ -11,7 +11,27 @@ export default class Versioning {
    * Return string representation of SDK version
    */
   static get sdkVersion(): string {
+<<<<<<< HEAD
+<<<<<<< HEAD
+<<<<<<< HEAD
+<<<<<<< HEAD
+<<<<<<< HEAD
     return '1.2.4';
+=======
+    return '1.1.18';
+>>>>>>> dfbc904... MeetingSessionPOSTLogger source and test
+=======
+    return '1.1.19';
+>>>>>>> a31f41e... Fix example code in the getting started guide
+=======
+    return '1.1.20';
+>>>>>>> 795512c... Add observer event for content sharing
+=======
+    return '1.1.21';
+>>>>>>> 7e534a9... Integrate MeetingSessionPOSTLogger to demo app
+=======
+    return '1.2.3';
+>>>>>>> 59c9ebd... Remove unimplemented callbacks
   }
 
   /**